Activity: Explore probability distributions
In this lab, you will learn how data across a dataset is distributed and identify data points that need further examination. You will be presented with a business scenario and a dataset to use as you explore the distribution. You will practice determining the type of probability distribution that fits the data, calculating z-score, and detecting outliers to understand how the data is distributed.

To complete this lab, you will open a Jupyter notebook and follow instructions to enter code and written responses where prompted.