Activity_Perform feature engineering
In this lab, you will perform feature engineering. You will be presented with a business scenario and a dataset to use as you explore the scenario. You will practice using functions in Python to identify meaningful variables, perform necessary transformations and engineer new features.

Data Dictionary
This activity uses a dataset called
nba-players.csv
. It represents 22 columns of United States National Basketball Association (NBA) player performance records per game for one season, including stats like average number of 3-point baskets made per game, and average number of offensive rebounds per game.

The dataset contains:

1,340 rows – each row is a different NBA player’s stats per game across one season

22 columns

Note: A “field goal” in basketball means the ball has gone through the hoop during play resulting in points scored. 2 and 3 points are scored during game play based on the distance the player is from the hoop. 1 point is scored during a “free throw,” representing a free shot at the basket due to a foul by the other team.