Activity: Build a Naive Bayes model
In this lab, you will build your own Naive Bayes model. You will be presented with a business scenario and a dataset to use as you explore using a decision tree model to make predictions for a target based on multiple features. You will practice modeling with Naive Bayes to quickly and more accurately make predictions about what could occur.

Data Dictionary
This activity uses a modified version of the NBA dataset used previously. The file is called
extracted-nba-players.csv
. It represents 11 columns of United States National Basketball Association (NBA) player performance records per game for one season, including stats like average number of 3-point baskets made per game, and average number of offensive rebounds per game.

The dataset contains:

1,340 rows – each row is a different NBA player’s stats per game across one season

11 columns

Note: A “field goal” in basketball means the ball has gone through the hoop during play resulting in points scored. 2 and 3 points are scored during game play based on the distance the player is from the hoop. 1 point is scored during a “free throw,” representing a free shot at the basket due to a foul by the other team.