Activity: Build a K-means model
In this lab, you will build a K-means model. You will be presented with a business scenario and a dataset to use as you explore the scenario. You will practice interpreting inertia and silhouette coefficients and decide on an optimal value of k based on these scores to determine how best to cluster the data.

Data dictionary
This activity uses a data set called
penguins.csv
. It represents data about three penguin species. The data was collected from Palmer Station in Antarctica.